Changing type of app basic vid and pid to Integer, Fixed constructor param type for TargetInfo and Conversion of param to Integer

+++ examples/tv-casting-app/android/App/app/src/main/jni/com/chip/casting/TvCastingApp.java
@@ -332,7 +332,7 @@ public class TvCastingApp {
 
   public native boolean applicationBasic_readVendorID(
       ContentApp contentApp,
-      SuccessCallback<Short> readSuccessHandler,
+      SuccessCallback<Integer> readSuccessHandler,
       FailureCallback readFailureHandler);
 
   public native boolean applicationBasic_readApplicationName(
@@ -342,7 +342,7 @@ public class TvCastingApp {
 
   public native boolean applicationBasic_readProductID(
       ContentApp contentApp,
-      SuccessCallback<Short> readSuccessHandler,
+      SuccessCallback<Integer> readSuccessHandler,
       FailureCallback readFailureHandler);
 
   public native boolean applicationBasic_readApplicationVersion(
